src/backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
dnspython==2.4.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-ldap==3.4.3
pytz==2023.3.post1
sqlparse==0.4.4
